chore: order csproj item groups for inclusion
Placed the production-dependencies ItemGroup above the test-dependencies
ItemGroup in the game csproj. Now when users execute "dotnet add
package", it will by default be added to the production-dependencies
group. This avoids production dependencies being added to a group that
is compiled out in non-test builds, causing unforeseen build errors
when building without tests.